some fix

+++ src/Pages/TakenFeatures/TakenFeatures.js
@@ -6,6 +6,9 @@ import { collection, getDocs, doc, deleteDoc } from "firebase/firestore";
 import { FaTrash } from "react-icons/fa";
 import Features from '../Features/Features';
 import { Link } from "react-router-dom"
+import { MdArrowBack } from "react-icons/md";
+import swal from 'sweetalert';
+import './TakenFeatures.css'
 
 const TakenFeatures = () => {
     const { id } = useParams()
@@ -13,15 +16,23 @@ const TakenFeatures = () => {
     const [profile, setProfile] = useState({})
     const [user, loading, error] = useAuthState(auth)
     const [dataChecked, setDataChecked] = useState(true)
+    const navigate = useNavigate()
     const deleteDocument = (docid) => {
-        const DeletedocRef = doc(db, `Users/${user?.uid}/ChildList/child1/data/${id}/${id}/${docid}`);
-        deleteDoc(DeletedocRef)
-            .then(() => {
-                const newData=data.filter(doc => doc.docId !== docid);
-                setData(newData)
-            })
-            .catch(error => {
-                console.log(error);
+        swal("Are you sure?", {
+            buttons: ["Cancel", true],
+        })
+            .then((res) => {
+                if (res === true) {
+                    const DeletedocRef = doc(db, `Users/${user?.uid}/ChildList/child1/data/${id}/${id}/${docid}`);
+                    deleteDoc(DeletedocRef)
+                        .then(() => {
+                            const newData = data.filter(doc => doc.docId !== docid);
+                            setData(newData)
+                        })
+                        .catch(error => {
+                            console.log(error);
+                        })
+                }
             })
     }
     if (loading) {
@@ -57,7 +68,7 @@ const TakenFeatures = () => {
                     <path d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z" fill="currentFill" />
                 </svg>
             </div>
-
+            
         }
         else {
 
@@ -69,16 +80,23 @@ const TakenFeatures = () => {
 
 
 
-    console.log(data)
 
     if (data.length > 0) {
         return (
             <>
+                <div className="bg-primary relative h-14 flex justify-center items-center">
+                    <p className="absolute left-6 font-medium cursor-pointer" onClick={() => navigate('/')}><MdArrowBack color='#fff' fontSize={25} /></p>
+                    <p className="text-white font-medium text-lg">
+                        {id == "emergencycont" && "Emergency Contacts"}
+                        {id == "mediccond" && "Medical Conditions"}
+                        {id !== "emergencycont" && id !== "mediccond" && id.charAt(0).toUpperCase() + id.slice(1)}
+
+                    </p>
+                </div>
                 <div className=' relative' style={{ height: "80vh" }}>
 
                     {
                         data?.map((data) => {
-                            console.log(data)
                             return <div className=' w-11/12 mx-auto bg-white p-5 rounded-xl mt-5'>
                                 <h1 className="font-medium">
                                     {
